+++ shared/tsconfig.json
@@ -1,9 +1,10 @@
-{
+{"extends": "../tsconfig.base.json",
   "compilerOptions": {
     "composite": true,
     "declaration": true,
     "declarationMap": true,
     "noEmit": false,
+    "rootDir": ".",
     "outDir": "dist",
     "module": "ESNext",
     "target": "ES2020",
@@ -12,6 +13,9 @@
     "moduleResolution": "node",
     "skipLibCheck": true
   },
-  "include": ["./**/*.ts"],
+  "include": [
+    "interface/**/*.ts",
+    "utils/**/*.ts"
+  ],
   "exclude": ["node_modules"]
 }
